set height on new publishes to current height in Claim db
