+++ test/end-to-end/end-to-end.test.js
@@ -1,3 +1,5 @@
+require('../module-alias-boilerplate.js');
+
 const chai = require('chai');
 const expect = chai.expect;
 const chaiHttp = require('chai-http');
