+++ test/module-alias-boilerplate.js
@@ -1,0 +1,4 @@
+// set up aliases
+const moduleAlias = require('module-alias');
+const customAliases = require('../utils/createModuleAliases.js')();
+moduleAlias.addAliases(customAliases);
